Issue template must not apply organisation wide
Since this template was added in the .github repository it applied to
all repositories in the CHERI Alliance. This is undesirable. Instead
I've moved this template to the RepMgmtWg as you can see in this PR:
https://github.com/CHERI-Alliance/RepoMgmtWG/pull/12
This PR removes that issue template from here and updates the reference
in the repository creation rules document.